fix(logo, header): implement dynamic color change for Logo and move Header component
