+++ frontend/src/components/Layout/Layout.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 
-import Header from '../Header/Header';
+import Header from '../sections/Header/Header';
 import Footer from '../sections/Footer/Footer';
 
 const Layout = () => {
