+++ frontend/src/components/Navigation/Navigation.module.css
@@ -4,11 +4,11 @@
   align-items: center;
 }
 
-/* @media screen and (min-width: 768px) {
+@media screen and (min-width: 768px) {
   .nav {
     gap: 177px;
   }
-} */
+}
 
 @media screen and (min-width: 1440px) {
   .nav {
@@ -72,4 +72,12 @@
 .darkActive {
   border: 1px solid var(--white);
   border-radius: 30px;
+}
+
+.logoLight {
+  fill: var(--black);
+}
+
+.logoDark {
+  fill: var(--white);
 }
