+++ frontend/src/components/sections/Header/Header.module.css
@@ -31,7 +31,6 @@
   justify-content: space-between;
   align-items: center;
   gap: 24px;
-  margin: 8px 8px 0 8px;
   padding: 16px;
   border-top-left-radius: 20px;
   border-top-right-radius: 20px;
@@ -40,7 +39,6 @@
 
 @media screen and (min-width: 768px) {
   .headerWrapper {
-    margin: 16px 16px 0 16px;
     padding: 20px 32px;
     border-top-left-radius: 30px;
     border-top-right-radius: 30px;
@@ -49,7 +47,6 @@
 
 @media screen and (min-width: 1440px) {
   .headerWrapper {
-    margin: 20px 20px 0 20px;
     padding: 20px 60px;
   }
 }
